SBMLWriter
Notes:
There are several likely issues with the XML writing. This is mostly
pseudocode to get an idea of how an XML writer will work for MOST. Just
need to gain a better familiarity with the javax.xml library and will be
able to correct the issues then.

Additionally, there are a lot of redundancies which can be handled by
creating various super-classes. To be taken care of around testing to
allow for changes to be more easily made.

More information about XML reading and writing can be found at:
http://www.vogella.com/articles/JavaXML/article.html

More information about attribute adding can be found at:

http://www.java2s.com/Code/Java/JDK-6/UsingXMLEventFactorytocreatexmldocument.htm